LPS-88444 Move initial upgrade step back to ModuleApplicationContextExtender, in preparation for reverting LPS-85683; was initially moved by LPS-86448 (97e49fe6)
