+++ modules/apps/portal/portal-spring-extender-impl/src/main/java/com/liferay/portal/spring/extender/internal/context/ModuleApplicationContextExtender.java
@@ -19,16 +19,26 @@ import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.configuration.Configuration;
+import com.liferay.portal.kernel.dao.db.DB;
+import com.liferay.portal.kernel.dao.db.DBContext;
+import com.liferay.portal.kernel.dao.db.DBManager;
+import com.liferay.portal.kernel.dao.db.DBProcessContext;
 import com.liferay.portal.kernel.io.unsync.UnsyncBufferedReader;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Release;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
+import com.liferay.portal.kernel.upgrade.UpgradeException;
+import com.liferay.portal.kernel.upgrade.UpgradeStep;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.HashMapDictionary;
 import com.liferay.portal.kernel.util.InfrastructureUtil;
+import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xml.SAXReaderUtil;
 import com.liferay.portal.spring.extender.internal.configuration.ConfigurationUtil;
 
+import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
@@ -45,6 +55,7 @@ import org.apache.felix.utils.extender.Extension;
 
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceRegistration;
 import org.osgi.framework.wiring.BundleWiring;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
@@ -124,6 +135,99 @@ public class ModuleApplicationContextExtender extends AbstractExtender {
 	private static final Log _log = LogFactoryUtil.getLog(
 		ModuleApplicationContextExtender.class);
 
+	private static class InitialUpgradeStep implements UpgradeStep {
+
+		@Override
+		public String toString() {
+			return "Initial Database Creation";
+		}
+
+		@Override
+		public void upgrade(DBProcessContext dbProcessContext)
+			throws UpgradeException {
+
+			DBContext dbContext = dbProcessContext.getDBContext();
+
+			DBManager dbManager = dbContext.getDBManager();
+
+			DB db = dbManager.getDB();
+
+			String tablesSQL = _getSQLTemplateString("tables.sql");
+			String sequencesSQL = _getSQLTemplateString("sequences.sql");
+			String indexesSQL = _getSQLTemplateString("indexes.sql");
+
+			if (tablesSQL != null) {
+				try {
+					db.runSQLTemplateString(tablesSQL, true, true);
+				}
+				catch (Exception e) {
+					throw new UpgradeException(
+						StringBundler.concat(
+							"Bundle ", _bundle,
+							" has invalid content in tables.sql:\n", tablesSQL),
+						e);
+				}
+			}
+
+			if (sequencesSQL != null) {
+				try {
+					db.runSQLTemplateString(sequencesSQL, true, true);
+				}
+				catch (Exception e) {
+					throw new UpgradeException(
+						StringBundler.concat(
+							"Bundle ", _bundle,
+							" has invalid content in sequences.sql:\n",
+							sequencesSQL),
+						e);
+				}
+			}
+
+			if (indexesSQL != null) {
+				try {
+					db.runSQLTemplateString(indexesSQL, true, true);
+				}
+				catch (Exception e) {
+					throw new UpgradeException(
+						StringBundler.concat(
+							"Bundle ", _bundle,
+							" has invalid content in indexes.sql:\n",
+							indexesSQL),
+						e);
+				}
+			}
+		}
+
+		private InitialUpgradeStep(Bundle bundle) {
+			_bundle = bundle;
+		}
+
+		private String _getSQLTemplateString(String templateName)
+			throws UpgradeException {
+
+			URL resource = _bundle.getResource("/META-INF/sql/" + templateName);
+
+			if (resource == null) {
+				if (_log.isDebugEnabled()) {
+					_log.debug("Unable to locate SQL template " + templateName);
+				}
+
+				return null;
+			}
+
+			try (InputStream inputStream = resource.openStream()) {
+				return StringUtil.read(inputStream);
+			}
+			catch (IOException ioe) {
+				throw new UpgradeException(
+					"Unable to read SQL template " + templateName, ioe);
+			}
+		}
+
+		private final Bundle _bundle;
+
+	}
+
 	private class ModuleApplicationContextExtension implements Extension {
 
 		public ModuleApplicationContextExtension(Bundle bundle) {
@@ -135,6 +239,10 @@ public class ModuleApplicationContextExtender extends AbstractExtender {
 
 		@Override
 		public void destroy() throws Exception {
+			if (_upgradeStepServiceRegistration != null) {
+				_upgradeStepServiceRegistration.unregister();
+			}
+
 			if (_component != null) {
 				_dependencyManager.remove(_component);
 			}
@@ -189,6 +297,9 @@ public class ModuleApplicationContextExtender extends AbstractExtender {
 			}
 
 			_dependencyManager.add(_component);
+
+			_upgradeStepServiceRegistration = _processInitialUpgrade(
+				bundleContext);
 		}
 
 		private void _generateConfigurationDependency(
@@ -224,6 +335,45 @@ public class ModuleApplicationContextExtender extends AbstractExtender {
 					_bundle.getVersion(), "))"));
 
 			_component.add(serviceDependency);
+		}
+
+		private ServiceRegistration<UpgradeStep> _processInitialUpgrade(
+			BundleContext bundleContext) {
+
+			Dictionary<String, String> headers = _bundle.getHeaders(
+				StringPool.BLANK);
+
+			String upgradeToSchemaVersion = GetterUtil.getString(
+				headers.get("Liferay-Require-SchemaVersion"),
+				headers.get("Bundle-Version"));
+
+			Dictionary<String, Object> properties = new HashMapDictionary<>();
+
+			BundleWiring bundleWiring = _bundle.adapt(BundleWiring.class);
+
+			ClassLoader classLoader = bundleWiring.getClassLoader();
+
+			Configuration configuration = ConfigurationUtil.getConfiguration(
+				classLoader, "service");
+
+			if (configuration != null) {
+				String buildNumber = configuration.get("build.number");
+
+				if (buildNumber != null) {
+					properties.put("build.number", buildNumber);
+				}
+			}
+
+			properties.put("upgrade.initial.database.creation", "true");
+
+			properties.put(
+				"upgrade.bundle.symbolic.name", _bundle.getSymbolicName());
+			properties.put("upgrade.db.type", "any");
+			properties.put("upgrade.from.schema.version", "0.0.0");
+			properties.put("upgrade.to.schema.version", upgradeToSchemaVersion);
+
+			return bundleContext.registerService(
+				UpgradeStep.class, new InitialUpgradeStep(_bundle), properties);
 		}
 
 		private void _processServiceReferences(ClassLoader classLoader)
@@ -274,6 +424,8 @@ public class ModuleApplicationContextExtender extends AbstractExtender {
 		private final Bundle _bundle;
 		private org.apache.felix.dm.Component _component;
 		private final DependencyManager _dependencyManager;
+		private ServiceRegistration<UpgradeStep>
+			_upgradeStepServiceRegistration;
 
 	}
